Update Parse/TryParse model
