+++ src/enumeration/Enumeration/Enumeration/Enumeration.ValueSet.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+using System.Collections.Generic;
+
+namespace System
+{
+    partial class Enumeration<TEnumeration>
+    {
+        private static IReadOnlySet<TEnumeration> ValueSet => EnumerationSet.Value;
+    }
+}
